requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
azure-ai-ml==1.29.0
azure-common==1.1.28
azure-core==1.35.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.0
azure-mgmt-core==1.6.0
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b42
azure-storage-blob==12.26.0
azure-storage-file-datalake==12.21.0
azure-storage-file-share==12.22.0
azureml-dataprep==4.10.6
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.17.12
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.3
cryptography==40.0.2
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dotnetcore2==3.1.23
exceptiongroup==1.3.0
executing==2.2.1
fixedint==0.1.6
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
mltable==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-urllib==0.58b0
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
pandas==1.5.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==9.0.0
pycparser==2.23
pydash==8.0.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
stack-data==0.6.3
strictyaml==1.7.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.3
zipp==3.23.0
